+++ src/main/kotlin/pl/diet/company/dietapp/product/repository/ProductRepository.kt
@@ -1,9 +1,9 @@
-package pl.diet.company.dietapp.repository
+package pl.diet.company.dietapp.product.repository
 
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.data.mongodb.repository.Query
 import org.springframework.stereotype.Component
-import pl.diet.company.dietapp.domain.Product
+import pl.diet.company.dietapp.product.domain.Product
 
 @Component
 interface ProductRepository : MongoRepository<Product, Long> {
