+++ src/main/kotlin/pl/diet/company/dietapp/registration/controller/RegistrationController.kt
@@ -1,0 +1,4 @@
+package pl.diet.company.dietapp.registration.controller
+
+class RegistrationController {
+}
