+++ src/main/kotlin/pl/diet/company/dietapp/security/UserRegistartion.kt
@@ -1,0 +1,4 @@
+package pl.diet.company.dietapp.security
+
+class UserRegistartion {
+}
