+++ src/main/kotlin/pl/diet/company/dietapp/security/filter/JwtRequestFilter.kt
@@ -1,4 +1,4 @@
-package pl.diet.company.dietapp.filter
+package pl.diet.company.dietapp.security.filter
 
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
@@ -6,8 +6,8 @@ import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
 import org.springframework.stereotype.Component
 import org.springframework.web.filter.OncePerRequestFilter
-import pl.diet.company.dietapp.service.DietAppUserService
-import pl.diet.company.dietapp.util.JwtUtil
+import pl.diet.company.dietapp.security.service.DietAppUserService
+import pl.diet.company.dietapp.security.util.JwtUtil
 import javax.servlet.FilterChain
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
