+++ src/test/kotlin/pl/diet/company/dietapp/TestProductRemove.kt
@@ -6,7 +6,7 @@ import org.springframework.boot.test.autoconfigure.data.mongo.AutoConfigureDataM
 import org.springframework.test.context.junit4.SpringRunner
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-import pl.diet.company.dietapp.util.TestBase
+import pl.diet.company.dietapp.security.util.TestBase
 
 @AutoConfigureDataMongo
 @RunWith(SpringRunner::class)
