+++ src/test/kotlin/pl/diet/company/dietapp/builder/ProductBuilder.kt
@@ -1,9 +1,9 @@
 package pl.diet.company.dietapp.builder
 
-import pl.diet.company.dietapp.domain.ProductCompositionDescription
-import pl.diet.company.dietapp.domain.PositiveMeasurement
-import pl.diet.company.dietapp.domain.Money
-import pl.diet.company.dietapp.domain.ProductRequest
+import pl.diet.company.dietapp.product.domain.ProductCompositionDescription
+import pl.diet.company.dietapp.product.domain.PositiveMeasurement
+import pl.diet.company.dietapp.product.domain.Money
+import pl.diet.company.dietapp.product.domain.ProductRequest
 import java.math.BigDecimal
 
 class ProductBuilder {
